// src/models/types/personal-data.ts
export type PersonalData = {
    ide: number;
    name: string;
    active: boolean;
    position?: string;
    area?: string;
    birthDate?: string;
    birthPlace?: string;
    birthState?: string;
    curp?: string;
    rfc?: string;
    militaryCard?: string;
    issstep?: string;
    gender?: string;
    maritalStatus?: string;
    phone?: string;
    mobile?: string;
    email?: string;
    institutionalMail?: string;
    address?: string;
    addressNumber?: string;
    colony?: string;
    population?: string;
    postalCode?: string;
    photo?: string;
    degree?: string;
    title?: string;
    titleComplete?: string;
}